Another test case and perf counter.
For the 'resume a bunch of broken txns' case.
It is definitely still O(N^2) due to calling rescan() for each
transaction, which then does a batch reload of all the other
transactions.

The bulk of the remaining time seems to be just in deserializing
the database objects into go objects. (for N=1000 we do 999 rescans,
and preload 997,000 documents.)

Preload is currently broken for this case because it loads all of
the transactions for a document but only flags the one transaction
it recurses into as 'seen' which causes it to then preload all but that
one seen transaction again.