Preload was also becoming N^2.
If you had lots of "prepared" documents, we would preload the transactions to
work on, and then recurse into the first transaction, which would then cause us to
preload all transactions affecting that document again.

Passing the preloaded map into recurse allows us to just reuse the objects. It changes

N    time   id()calls txn preload count rescan count
1000    6s    503502     997001          500499
2000   27s   2007002    3994001         2000999
4000  104s   8014002   15988001         8001999

 into

N    time   id()calls txn preload count rescan count
1000  1.4s    503502     1000            500499
2000  4.6s   2007002     2000           2000999
4000 17.4s   8014002     4000           8001999

We are now calling preload for the right number of documents, and
this has significantly improved the runtime cost. We are still triggering
O(N^2) id() calls in tokensWithIds() because rescan() is not able to
reuse the existing Queue (when we get around to rescan() the size of the
queue has changed because the previous transaction has been applied.)
It seems to concretely be (N*N-1)/2+N-1 rescanned converted txn ids.

Trying to lazily cache txn.id() just makes it worse, because we add the
overhead of checking that if the Id is available (and probably messes with
inlining?).

The current overhead is out of 15.5s to Run() the txn, pprof only samples about
8.6s of it. Of that time, 4.7s is in 'rescan' split between 3.2s of FindId().One()
actually reloading the document, and 1.26s in tokensWithIds calling .id() lots
of times. The only way to avoid that would be to have rescan not actually reload
the txn from the database, but that doesn't seem valid.